requirements.txt
Pillow==9.4.0
Unidecode~=1.3.6
